+++ programs/multisig-lite/src/lib.rs
@@ -384,6 +384,57 @@ pub mod multisig_lite {
     ///
     /// It's restricted one multisig account to each funder Pubkey,
     /// as it's used for the multisig PDA address generation.
+    ///
+    /// # Examples
+    ///
+    /// Here is how you create a multisig account on devnet:
+    ///
+    /// ```no_run
+    /// use std::rc::Rc;
+    ///
+    /// use solana_sdk::commitment_config::CommitmentConfig;
+    /// use solana_sdk::pubkey::Pubkey;
+    /// use solana_sdk::signature::read_keypair_file;
+    /// use solana_sdk::signer::Signer;
+    /// use solana_sdk::system_program;
+    ///
+    /// use anchor_client::{Client, Cluster};
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let funder = Rc::new(read_keypair_file(
+    ///     shellexpand::tilde("~/.config/solana/id.json").as_ref(),
+    /// )?);
+    /// let url = Cluster::Devnet;
+    /// let opts = CommitmentConfig::processed();
+    /// let pid = multisig_lite::id();
+    /// let program = Client::new_with_options(url, funder.clone(), opts).program(pid);
+    ///
+    /// // Gets the PDAs.
+    /// let (state_pda, state_bump) =
+    ///     Pubkey::find_program_address(&[b"state", funder.pubkey().as_ref()], &pid);
+    /// let (fund_pda, fund_bump) = Pubkey::find_program_address(&[b"fund", state_pda.as_ref()], &pid);
+    ///
+    /// // Creates a multisig account.
+    /// let sig = program
+    ///     .request()
+    ///     .accounts(multisig_lite::accounts::Create {
+    ///         funder: funder.pubkey(),
+    ///         state: state_pda,
+    ///         fund: fund_pda,
+    ///         system_program: system_program::id(),
+    ///     })
+    ///     .args(multisig_lite::instruction::Create {
+    ///         m: 2, // m as in m/n.
+    ///         signers: vec![funder.pubkey(), Pubkey::new_unique(), Pubkey::new_unique()],
+    ///         q: 10, // queue limit.
+    ///         _state_bump: state_bump,
+    ///         fund_bump,
+    ///     })
+    ///     .signer(funder.as_ref())
+    ///     .send()?;
+    /// # Ok(())
+    /// # }
+    /// ```
     #[allow(clippy::result_large_err)]
     pub fn create(
         ctx: Context<Create>,